requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.6.0
    # via pydantic
ansible==9.5.1
    # via pipeformer (pyproject.toml)
ansible-core==2.16.6
    # via ansible
anyio==4.3.0
    # via httpx
bcrypt==4.1.2
    # via paramiko
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
click==8.1.7
    # via pipeformer (pyproject.toml)
cryptography==42.0.5
    # via
    #   ansible-core
    #   paramiko
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via pipeformer (pyproject.toml)
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via pipeformer (pyproject.toml)
idna==3.7
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via ansible-core
markupsafe==2.1.5
    # via jinja2
packaging==24.0
    # via
    #   ansible-core
    #   pytest
paramiko==3.4.0
    # via pipeformer (pyproject.toml)
pluggy==1.5.0
    # via pytest
pycparser==2.22
    # via cffi
pydantic==2.7.0
    # via pipeformer (pyproject.toml)
pydantic-core==2.18.1
    # via pydantic
pynacl==1.5.0
    # via paramiko
pytest==8.2.0
    # via pipeformer (pyproject.toml)
pyyaml==6.0.1
    # via
    #   ansible-core
    #   pipeformer (pyproject.toml)
resolvelib==1.0.1
    # via ansible-core
sentry-sdk==2.0.1
    # via pipeformer (pyproject.toml)
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via sentry-sdk
